esp32 support, nsh console running

+++ platforms/nuttx/src/px4/common/SerialImpl.cpp
@@ -129,7 +129,7 @@ bool SerialImpl::configure()
 
 	default:
 		speed = _baudrate;
-		PX4_WARN("Using non-standard baudrate: %lu", _baudrate);
+		PX4_WARN("Using non-standard baudrate: %lu", (long unsigned)_baudrate);
 		break;
 	}
